finished the numeric input bar unit tests

+++ src/BlazorRamp.Inputs/Sandbox.Contracts/Models/AllModels.cs
@@ -32,7 +32,7 @@ public record class ContactDto
     public DateOnly CompareDOB     { get; set; } = default!;
     public string   Email          { get; set; } = default!;
     public string?  NullableMobile { get; set; } 
-    public int?     NullableAge    { get; set; }
+    public decimal?     NullableAge    { get; set; }
     public int      Age            { get; set; }
     public string   Password       { get; set; } = default!;
     public List<string> Entries { get; set; } = [];
